+++ controller/reportStockTelemetry.ts
@@ -50,13 +50,13 @@ export async function main() {
     const globalTelemetries = await sskts.service.report.telemetry.searchGlobalStock({
         measuredFrom: measuredFrom.toDate(),
         measuredThrough: dateNowByUnitTime.toDate()
-    })(telemetryRepo);
+    })({ telemetry: telemetryRepo });
     debug('globalTelemetries length:', globalTelemetries.length);
 
     const sellerTelemetries = await sskts.service.report.telemetry.searchSellerStock({
         measuredFrom: measuredFrom.toDate(),
         measuredThrough: dateNowByUnitTime.toDate()
-    })(telemetryRepo);
+    })({ telemetry: telemetryRepo });
     debug('sellerTelemetries length:', sellerTelemetries.length);
 
     debug('diconnecting mongo...');